[Rename] input 폴더 이름 수정(2023/09/06)
