+++ src/components/input/NicknameInput.jsx
